# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Open-Kanban API
  description: |
    Open-Kanban is a self-hosted kanban board API server.
    
    ## Authentication
    
    This API uses Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    Some endpoints (marked as public) do not require authentication.
    
    ## Signature Verification (IMPORTANT SECURITY INFORMATION)
    
    Protected endpoints support HMAC-SHA256 signature verification for enhanced security. However, **signature verification is DISABLED by default**.
    
    ### Security Warning
    
    When signature verification is disabled (the default), protected endpoints will accept requests without signature verification, even if they are marked as requiring signature verification in this documentation.
    
    ### To Enable Signature Verification
    
    Set the following environment variables:
    
    1. `SIGNATURE_ENABLED=1` - Enable signature verification
    2. `SIGNATURE_SECRETS=access-key-1:secret-key-1,access-key-2:secret-key-2` - Configure access keys and their corresponding secrets (comma-separated key:value pairs)
    
    ### How Signature Verification Works
    
    When enabled, protected endpoints require the following headers:
    
    - `X-Signature`: HMAC-SHA256 signature (hex-encoded)
    - `X-Timestamp`: Unix timestamp of the request (must be within 5 minutes of server time)
...
    
    ### Signature Generation
    
...
    
...
servers:
...
paths:
...
  /columns:
    get:
      summary: Get columns
      description: Public endpoint to get columns, optionally filtered by board
      operationId: getColumns
      tags:
        - Columns
      parameters:
        - name: boardId
          in: query
          schema:
            type: string
        - name: positions
          in: query
          description: Comma-separated list of positions
          schema:
            type: string
      responses:
        '200':
          description: List of columns with tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Column'
    post:
      summary: Create a column
      description: Create a new column in a board
      operationId: createColumn
      tags:
        - Columns
      security:
        - BearerAuth: []
        - SignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColumnRequest'
      responses:
        '200':
          description: Column created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Column'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      summary: Update a column
      description: Update column properties
      operationId: updateColumn
      tags:
        - Columns
      security:
        - BearerAuth: []
        - SignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateColumnRequest'
      responses:
        '200':
          description: Column updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      summary: Delete a column
      description: Delete a column by ID
      operationId: deleteColumn
      tags:
        - Columns
      security:
        - BearerAuth: []
        - SignatureAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Column deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /tasks:
    get:
      summary: Get tasks
      description: Public endpoint to get tasks with optional filters
      operationId: getTasks
      tags:
        - Tasks
      parameters:
        - name: columnId
          in: query
          schema:
            type: string
        - name: boardId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of tasks with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  pageCount:
                    type: integer
    post:
      summary: Create a task
      description: Create a new task in a column
      operationId: createTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
        - SignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Rate limit exceeded
...
  /auth/permissions:
    get:
      summary: Get permissions
      description: Get all permissions
      operationId: getPermissions
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - SignatureAuth: []
      responses:
        '200':
          description: List of permissions
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Set permission
      description: Create or update a permission
      operationId: setPermission
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - SignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Permission set
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Delete permission
      description: Delete a permission
      operationId: deletePermission
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - SignatureAuth: []
      responses:
        '200':
          description: Permission deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/permissions/columns:
    get:
      summary: Get column permissions
      description: Get permissions for columns
      operationId: getColumnPermissions
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - SignatureAuth: []
      responses:
        '200':
          description: Column permissions
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Set column permission
      description: Create or update column permission
      operationId: setColumnPermission
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - SignatureAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Column permission set
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Delete column permission
      description: Delete a column permission
      operationId: deleteColumnPermission
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - SignatureAuth: []
      responses:
        '200':
          description: Column permission deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature
  schemas:
...
    Column:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        position:
          type: integer
        color:
          type: string
        description:
          type: string
        ownerAgentId:
          type: string
        boardId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        agentConfig:
          type: object
          properties:
            agentTypes:
              type: array
              items:
                type: string
    CreateColumnRequest:
      type: object
      required:
        - name
        - boardId
      properties:
        name:
          type: string
          maxLength: 100
        status:
          type: string
          maxLength: 50
        position:
          type: integer
        color:
          type: string
          maxLength: 20
        boardId:
          type: string
    UpdateColumnRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 100
        status:
          type: string
          maxLength: 50
        position:
          type: integer
        color:
          type: string
          maxLength: 20
        description:
          type: string
          maxLength: 500
        ownerAgentId:
          type: string
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
        meta:
          type: object
        columnId:
          type: string
        position:
          type: integer
        published:
          type: boolean
        archived:
          type: boolean
        archivedAt:
          type: string
          format: date-time
        agentId:
          type: string
        agentPrompt:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        commentCount:
          type: integer
        subtaskCount:
          type: integer
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        subtasks:
          type: array
          items:
            $ref: '#/components/schemas/Subtask'
    CreateTaskRequest:
      type: object
      required:
        - title
        - columnId
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
          maxLength: 100
        meta:
          type: object
        columnId:
          type: string
        position:
          type: integer
        published:
          type: boolean
        agentId:
          type: string
        agentPrompt:
          type: string
          maxLength: 2000
...
    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        author:
          type: string
        taskId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Subtask:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        completed:
          type: boolean
        taskId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not logged in
    Forbidden:
      description: Forbidden - User does not have permission
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: No permission
...